requirements.txt
streamlit==0.88.0
azure.cognitiveservices.speech==1.19.0
